Fixed compiling warning with gcc on --enable-64bit mode.
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@15860 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/common/utils.c
@@ -45,14 +45,14 @@ void WriteDump(FILE* fp, const void* buffer, size_t length)
 
 		if( (i%16) == 15 )
 		{
-			fprintf(fp, "%03X %s  %s\n", i/16, hex, ascii);
+			fprintf(fp, "%03X %s  %s\n", (unsigned int)(i/16), hex, ascii);
 		}
 	}
 
 	if( (i%16) != 0 )
 	{
 		ascii[i%16] = 0;
-		fprintf(fp, "%03X %-48s  %-16s\n", i/16, hex, ascii);
+		fprintf(fp, "%03X %-48s  %-16s\n", (unsigned int)(i/16), hex, ascii);
 	}
 }
